Move singleSignOnContext export to index, update tests
