+++ sso-kit-client/core/tsconfig.json
@@ -5,6 +5,5 @@
     "rootDir": "src",
     "outDir": "src"
   },
-  "include": ["src"],
   "exclude": ["test/**/*.ts"]
 }
