+++ sso-kit-client/lit/tsconfig.json
@@ -4,7 +4,6 @@
     "rootDir": "src",
     "outDir": "./lib/lit/src"
   },
-  "include": ["src"],
   "exclude": ["test/**/*.ts"],
   "references": [
     {
