+++ sso-kit-client/react/src/SingleSignOnContext.ts
@@ -206,7 +206,3 @@ export class SingleSignOnContext {
     }
   };
 }
-
-export default function singleSignOnContext() {
-  return new SingleSignOnContext(window.Vaadin.SingleSignOnData!);
-}
